requirements.txt
backports.functools-lru-cache==1.5
cloudpickle==1.2.1
cycler==0.10.0
decorator==4.4.0
future==0.17.1
h5py==2.9.0
kiwisolver==1.1.0
matplotlib==2.2.4
networkx==2.2
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
Pillow==6.1.0
pkg-resources==0.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
scikit-image==0.14.3
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.4
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.2
